refactor: разделение контроллера, удаление зависимостей
- лишние зависимости закомментированы
- контроллер auth разбит на registration. login, refreshTokens, test
