+++ build.gradle.kts
@@ -35,16 +35,16 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("io.grpc:grpc-services")
     implementation("net.devh:grpc-client-spring-boot-starter:2.15.0.RELEASE")
-    implementation("net.devh:grpc-server-spring-boot-starter:2.15.0.RELEASE")
+//    implementation("net.devh:grpc-server-spring-boot-starter:2.15.0.RELEASE")
 
     implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server")
     implementation("org.springframework.boot:spring-boot-starter-actuator")
     implementation("org.springframework.session:spring-session-jdbc")
-    implementation("org.mapstruct:mapstruct:1.5.5.Final")
+//    implementation("org.mapstruct:mapstruct:1.5.5.Final")
     implementation("io.jsonwebtoken:jjwt-api:0.11.5")
     implementation ("io.grpc:grpc-protobuf:1.73.0")
     implementation ("io.grpc:grpc-stub:1.73.0")
-    implementation ("com.google.protobuf:protobuf-java:3.25.1")
+//    implementation ("com.google.protobuf:protobuf-java:3.25.1")
     compileOnly("org.projectlombok:lombok")
     compileOnly ("org.apache.tomcat:annotations-api:6.0.53")
     runtimeOnly("org.postgresql:postgresql")
