--- src/main/java/io/github/pavelshe11/authmicro/api/constants/APIRoutes.java
@@ -1,18 +1,0 @@
-package io.github.pavelshe11.authmicro.api.constants;
-
-public class APIRoutes {
-    private APIRoutes() {
-    }
-
-    public static final String AUTH = "/auth";
-    public static final String API_VERSION = "/v1";
-
-    public static final String REGISTRATION = "/registration";
-
-    public static final String LOGIN = "/login";
-
-    public static final String CONFIRM = "/confirmEmail";
-    public static final String SEND_CODE = "/sendCodeEmail";
-
-    public static final String REFRESH_TOKEN = "/refreshToken";
-}
