+++ src/main/java/io/github/pavelshe11/authmicro/api/controllers/LoginController.java
@@ -1,0 +1,38 @@
+package io.github.pavelshe11.authmicro.api.controllers;
+
+import io.github.pavelshe11.authmicro.api.dto.requests.LoginConfirmRequestDto;
+import io.github.pavelshe11.authmicro.api.dto.requests.LoginRequestDto;
+import io.github.pavelshe11.authmicro.api.dto.responses.LoginConfirmResponseDto;
+import io.github.pavelshe11.authmicro.api.dto.responses.LoginResponseDto;
+import io.github.pavelshe11.authmicro.services.LoginService;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/auth/v1/login")
+public class LoginController {
+    private final LoginService loginService;
+
+    @PostMapping("/sendCodeEmail")
+    public LoginResponseDto sendLoginCode(
+            @Valid @RequestBody LoginRequestDto loginRequest
+    ) {
+        // TODO: send email
+        return loginService.login(loginRequest.getEmail());
+    }
+
+    @PostMapping("/confirmEmail")
+    public LoginConfirmResponseDto confirmLoginEmail(
+            @Valid @RequestBody LoginConfirmRequestDto loginConfirmRequest
+    ) {
+        return loginService.confirmLoginEmail(
+                loginConfirmRequest.getEmail(),
+                loginConfirmRequest.getCode()
+        );
+    }
+}
